backend/requirements.txt
Flask==2.3.3
Flask-RESTX==1.1.0
Flask-Migrate==4.0.4
Flask-JWT-Extended==4.4.4
Flask-Cors==3.0.10
SQLAlchemy==1.4.49
psycopg2-binary==2.9.6
python-dotenv==1.0.0
marshmallow==3.19.0
alembic==1.11.1
Werkzeug<3.0.0
